@@ -1,34 +1,45 @@
 {
-  "name": "opulence/io",
-  "description": "The Opulence IO library",
-  "keywords": [
-    "PHP",
-    "file system",
-    "io"
-  ],
-  "authors": [
-    {
-      "name": "David Young",
-      "email": "dave@opulencephp.com",
-      "role": "Architect"
-    }
-  ],
-  "license": "MIT",
-  "autoload": {
-    "psr-4": {
-      "Opulence\\IO\\": ""
+    "name": "aphiria/io",
+    "description": "The Aphiria input/output library",
+    "keywords": [
+        "PHP",
+        "io",
+        "input output",
+        "streams",
+        "files"
+    ],
+    "authors": [
+        {
+            "name": "David Young",
+            "email": "dave@aphiria.com",
+            "role": "Architect"
+        }
+    ],
+    "license": "MIT",
+    "autoload": {
+        "psr-4": {
+            "Aphiria\\IO\\": "src"
+        }
     },
-    "exclude-from-classmap": [
-        "/Tests/"
-    ]
-  },
-  "require": {
-    "php": "^7.3"
-  },
-  "extra": {
-    "branch-alias": {
-      "dev-master": "2.0.x-dev"
-    }
-  },
-  "minimum-stability": "dev"
+    "autoload-dev": {
+        "psr-4": {
+            "Aphiria\\IO\\Tests\\": "tests"
+        }
+    },
+    "require": {
+        "php": "^7.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~8.0"
+    },
+    "scripts": {
+        "lint-check": "php-cs-fixer fix --config=.php_cs -v --diff --dry-run",
+        "lint-fix": "php-cs-fixer fix --config=.php_cs -v --diff"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
 }
